+++ fluxo-kmp-conf/src/main/kotlin/KmpHierarchyTemplate.kt
@@ -14,15 +14,33 @@ import fluxo.conf.dsl.container.impl.KmpTargetContainerImpl.NonJvm.Native.Nix.Li
 import org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi
 import org.jetbrains.kotlin.gradle.plugin.KotlinHierarchyBuilder
 import org.jetbrains.kotlin.gradle.plugin.KotlinHierarchyTemplate
+import org.jetbrains.kotlin.gradle.plugin.KotlinPlatformType
+import org.jetbrains.kotlin.gradle.plugin.KotlinSourceSetTree
 
 /**
  * Extends the Kotlin hierarchy template with the `commonJvm`,
  * `nonJvm`, `commonJs` and other groups.
+ *
+ * @see org.jetbrains.kotlin.gradle.plugin.KotlinHierarchyTemplate.Templates.default
+ * @see org.jetbrains.kotlin.gradle.plugin.defaultKotlinHierarchyTemplate
  */
 @OptIn(ExperimentalKotlinGradlePluginApi::class)
 public val KotlinHierarchyTemplate.Templates.fluxoKmpConf: KotlinHierarchyBuilder.Root.() -> Unit
     get() = {
+        // References:
+        // https://github.com/coil-kt/coil/blob/5d8bbca/buildSrc/src/main/kotlin/coil3/hierarchyTemplate.kt#L11
+        // https://github.com/touchlab/KaMPKit/blob/3af2a9f/shared/build.gradle.kts#L52
+        // https://github.com/proto-at-block/bitkey/blob/15b7c44/app/gradle/build-logic/src/main/kotlin/build/wallet/gradle/logic/extensions/KotlinMultiplatformExtension.kt#L47
+        // https://github.com/pdvrieze/xmlutil/blob/663c419/project-plugins/src/main/kotlin/net/devrieze/gradle/ext/nativeTargets.kt#L72
+
+        /* natural hierarchy is only applied to default 'main'/'test' compilations (by default) */
+        withSourceSetTree(KotlinSourceSetTree.main, KotlinSourceSetTree.test)
+
+        /** @see org.jetbrains.kotlin.gradle.plugin.KotlinHierarchyBuilder.common */
         group("common") {
+            /* All compilations shall be added to the common group by default. */
+            withCompilations { true }
+
             group(COMMON_JVM) {
                 withJvm()
                 try {
@@ -31,6 +49,15 @@ public val KotlinHierarchyTemplate.Templates.fluxoKmpConf: KotlinHierarchyBuilde
                     // Fallback for old Kotlin
                     @Suppress("DEPRECATION")
                     withAndroid()
+                }
+
+                excludeCompilations {
+                    when (it.platformType) {
+                        KotlinPlatformType.jvm,
+                        KotlinPlatformType.common,
+                        KotlinPlatformType.androidJvm -> false
+                        else -> true
+                    }
                 }
             }
             group(NON_JVM) {
@@ -48,8 +75,12 @@ public val KotlinHierarchyTemplate.Templates.fluxoKmpConf: KotlinHierarchyBuilde
                     }
                 }
                 group(NATIVE) {
+                    withNative()
+
                     group(NIX) {
                         group(APPLE) {
+                            withApple()
+
                             withIos()
                             withMacos()
                             withTvos()